Web Application/requirements.txt
cryptography==3.4.8
Flask==2.0.1
PyMySQL==1.0.2
PyYAML==5.4.1
pandas==1.3.2
scipy==1.8.0
urllib3==1.26.6
numpy==1.21.2
requests==2.26.0

 
